Update to try and use scaling in icon load

+++ include/globals.h
@@ -259,7 +259,7 @@ extern ODLayerList     *g_pLayerList;
 extern int             g_navobjbackups;
 extern int             g_EdgePanSensitivity;
 extern int             g_InitialEdgePanSensitivity;
-extern int             g_IconDisplayScaleFactor;
+extern double          g_IconDisplayScaleFactor;
 
 extern wxString        *g_ODlocale;
 extern int             g_iLocaleDepth;
